+++ lib/src/api_request_client.dart
@@ -4,6 +4,7 @@ import 'package:api_request/src/interceptors/unauthenticated_interceptor.dart';
 import 'package:flutter/foundation.dart';
 
 import '../api_request.dart';
+import 'api_log_level.dart';
 
 /// A singleton HTTP client wrapper around Dio with automatic configuration.
 ///
@@ -142,7 +143,7 @@ class RequestClient {
       headers: ApiRequestOptions.instance!.defaultHeaders,
     );
 
-    if (!kReleaseMode && ApiRequestOptions.instance!.enableLog) {
+    if (!kReleaseMode && ApiRequestOptions.instance!.logLevel != ApiLogLevel.none) {
       addInterceptorOnce(ApiLogInterceptor());
     }